layout: add tabs in buy page
